Replace the welcome message with an ActorProducer
What : Replace the welcome message the server sends with an actor producer,
Why : We need to be able to send messages on the server-side even though the client does not send any request.
How : Use ActorProducer instead of Source.single(ByteString(welcomeMsg)). Also use Merge instead of Concat on the Graph implementation.

For adding ActorProducer to the stream :
http://doc.akka.io/docs/akka-stream-and-http-experimental/2.0.3/scala/stream-integrations.html#integrating-with-actors

Issue #7