Done : analysis of openblockchain by IBM
out of 40K lines of code, 10K lines of code is test code.
The block structure is similar to Bitcoin's, having previous block hash, ...
The block is stored using Facebook's RocksDB.
For managing certificates it is using SQLite
Smart contracts are written in go language, and it is called chaincode.
Chaincode calls GetState, and PutState to adjust the amount of money an account has.
It uses ECC, with P384 curve for creating and validating signatures.
Example of chaincode is here. https://github.com/openblockchain/obc-peer/blob/master/openchain/example/chaincode/chaincode_example02/chaincode_example02.go
Summarized the result of analysis in Korean on my facebook account.

https://www.facebook.com/kangmo.kim.1/posts/800882396684097