Add a dummy method appendBlock on BlockWriter.
- Append a block to the given disk block storage,
producing file record locators for the block as well as each transaction in the block.

- Why? When we put a block into disk block storage, we have to create an index by block hash.
We also have to create an index by transaction hash that points to each transaction in the written block.

- This is necessary to read a specific transaction by hash, to get unspent output using an out point.
(An out point points to an output of a transaction using transaction hash and output index. )

Issue : #21