Fix Issue : Transaction validation fails if we have more than one transaction input.
- What
The block 707 we got from bitcoin core by connecting to it does not have an unlocking script on an input of a transaction. ( Attached badblock707.txt)

The correct block data is attached as block707.txt, this was dumped by DumpChain utility by reading a block file written by the same Bitcoin core instance running on my local machine.

- Why
While we calculate the hash for the signature validation of a specific transaction input, we changed unlocking script of all other inputs to empty. But we did not revert the unlocking scripts back to the original one after the signature validation.

- How
1. Do no use var for the unlockingScript. use val.
2. Instead of shallow copying a transaction, copy the transaction case class and create a new transaction by replacing the locking scripts with ones for signature validation.
3. Do not change the original transaction.

Issue : #31