Fix issue : CheckMultiSig fails to verify signature format for a specific transcation.
Error code and message :
[[ErrorCode(invalid_signature_format)]message=ScriptOp:CheckSig]

- root cause
The transaction has a negative value for the first byte of the S value of the signature.

- solution
Looks like IsValidCanonicalEncoding, which checks if the first byte of S value of the signature was written after the above transaction was put into blockchain.

We will skip checking if the first byte of the S value of the signature is positive.

We have more evidence of allowing negative values for S value string, but not sure how bitcoin core implementation passes the validation of the above transaction.

from : https://github.com/bitcoin/bips/blob/master/bip-0062.mediawiki
Inherent ECDSA signature malleability ECDSA signatures themselves are already malleable: taking the negative of the number S inside (modulo the curve order) does not invalidate it.

Issue : #33