Fix issue-#36 : Unable to decode the following block after reading it from a record storage.
- Problem

Unable to decode the following block after reading it from a record storage.

- Root Cause
RecordFile.appendRecord was overwriting data if readRead was called.

- Solution
Check if we are at the end of the file when appendRecord was called.
If not, move to the end of the file.

Need to fix one more problem to close this issue. Check the comment.