Done : Analysis for block validation
The blocks are validated using rules that were active by the time the block was created. The function call starts from the ProcessMessage, which calls ProcessNewBlock, which calls ActivateBestChain, which calls ActivateBestChainStep, which calls ConnectTip, which calls ConnectBlock, which creates flags based on the time when the block was created.

Issue : #44