As an alternative solution to the net layer, use Netty 4 for peer to peer communication.
Akka Streams is too much complicated. And also we found an issue that some servers are not able to send messages to others according to our test.
We switched to Netty 4, as it is more reliable. It is used by many project, so we can troubleshoot issues we hit.

1. Implement full duplex p2p communication using SecureChatServer and SecureChatClient example of Netty 4.
2. Also added encryption support for channels between nodes.

Issue : #60