Instead of HashCalculator, use transaction.hash, blockHeader.hash in codes.
- what
Remove HashCalculator.
Currently, we call HashCalculator.blockHeaderHash(blockHeader) to get the header of the block hash.
Instead, change the code to blockHeader.hash .

- why
For better code readability.

- how
The BlockHeader class is in the proto layer, whereas HashCalculator is in the crypto layer.
Because the proto layer is below the crypto layer, the BlockHeader class can not use the hash calculating functions in the crypto layer.
Create a new class called BlockHeaderEx in the crypto layer extending BlockHeader class.
add hash method to the blockHeaderEx, and create implicit type converter from the BlockHeader to the BlockHeaderEx.

Issue : #63