An orphan block is not added to main chain permanently.
Problem :
An orphan block is not added to main chain permanently.
This is reproduced when running five nodes more than a day. As a result of this issue, a node is not able to catch up other nodes.

Root cause :
CoinMiner propagates the newly mined block A to peers, and then it calls chain.putBlock.
Before the putBlock finishes, another node mines a new block B as a child of the block A.
As the putBlock was not finished, the block B is put as an orphan.
Also, because CoinMiner does not takes care of orphans after the chain.putBlock succeeds, the block B is never put into the blockchain.

Solution :
Call chain.putBlock before propagating the block to the network.

Issue : #74