Update block height index for block reorganization
- problem

During the long haul test, when we create more than 20K blocks, we hit a ChainException, resulting in disconnection with peers.

- root cause

The ChainException happens while getting a block by height. The code assumes that a block exists for the given height, but the block does not exist for the given height, resulting a ChainException.
This happened because we did not update the block height index for the block reorganization case.

- solution

Update the block height index for the block reorganization.

Issue : #78