Update Transaction descriptors for attached/detached transactions
- What
Whenever a transaction is attached or detached, we need to update the transaction descriptor of it.
1. When a transaction is attached, need to put the transaction descriptor with the transaction locator in a block on a blkNNNNN.dat file.
2. When a transaction is detached, need to del the transaction descriptor.

- Why
When the block reorganization runs, the same transaction may exist in a newly attached block. We need to delete the old transaction descriptor that has the transaction locator that points to the old detached block.

When a block is attached, each transaction in the block will put a transaction descriptor, whose locator points to the transaction location in the newly attached block.

- How
1. In Blockchain.putBlock, use BlockMagnet instead of updating indexes for the block and transactions in it.
2. Change DiskBlockStorage.putBlock to return the list of transaction locators. These locators are used by test codes only.
3. BlockWriter.getTxLocators returns transaction locators of a block without writing the block.
4. Write a test case to make sure transaction locators returned by BlockWriter.getTxLocators  and DiskBlockStorage.putBlock are identical.
5. Use  BlockWriter.getTxLocators to get the transaction locators of a newly attached block to put transaction descriptors of transactions in the block.

Also, removed Cassandra integration. We will implement our own data distribution instead of using Cassandra.