Separate the spent inpoints for transaction pool
- What

Currently, the spending in-points of a transaction in the transaction pool is stored in the transaction descriptor.
Move the spending in-points to the transaction pool, and separate the transaction descriptor and transaction pool completely.

- Why

While mining a block, we need to have a temporary mapping of transactions to spending in-points for each transaction to add to the new block to check transaction dependency. But because the spending in-points of transactions in (1) blocks and (2) the transaction pool are mixed in the transaction descriptor, it is hard to have such a temporary mapping.

- How

Move spending in-points to the transaction pool.

Issue : #82