Fix issue : Assertion hit on the BlockProcessor while processing the same block by two threads.
Root cause :
Two threads can run at the same time to process the same block, if the same block was receved from two different nodes.
An orphan block might have been removed by another thread while accepting chilren of a newly added block, but there was an assertion that an orphan block at the begining of BlockProcessor was still an orphan. This assumption is false.

Solution :
If a block is not an orphan anymore, do not accept it as a new child.