Adding a transaction takes longer with many addresses
- Problem

As time goes on, addresses are added to the wallet. As the number of addresses in the wallet grows,
the elapsed time for the following cases grows as well.
1. adding a transaction
2. register/unregister transaction on the wallet during block reorganization

- Root cause

adding a transaction registers the transaction on the wallet.
registerTransaction, unregisterTransaction on Wallet loops for each OutputOwnership which is the super class for the CoinAddress and ParsedPubKeyScript.

- Solution

Modify registerTransaction, unregisterTransaction to search RocksDB index instead of looping each OutputOwnership if possible.

Issue : #86