The response of getrawtransaction does not include following fields. Add them as the response. ``` //  blockhash     : Option[Hash],   // "00000000103e0091b7d27e5dc744a305108f0c752be249893c749e19c1c82317",   // If the transaction has been included in a block on the local best block chain,   // this is how many confirmations it has. Otherwise, this is 0 //  confirmations : Long,   // 88192,   // If the transaction has been included in a block on the local best block chain,   // this is the block header time of that block (may be in the future) //  time          : Option[Long],   // 1398734825,   // This field is currently identical to the time field described above //  blocktime     : Option[Long]    // 1398734825 ```
Need these fields to check if the transaction is included in a block.

Add a new field, blockHeightOption in TransactionDescriptor. When a transaction is included in a block, the blockHeightOption is set to the height of the block in the local best blockchain.
When getrawtransaction RPC is invoked, construct the above fields based on the block at the given height.

Issue : #89