Fix issue : Sendrawtransaction hangs while block reorganization is in progress.
Root cause

Block reorganization attaches/detaches transactions in blocks. Sendrawtransaction has to wait for the block reorganization, as the transaction could be conflicting with one in blocks attaches/detached during the reorg.

Solution

On each peer, always accept blocks in the same order. Do not process a block, if another block was already attached on the same previous block.

Issue : #95