A node doesn't accept blocks anymore after running tests a while
Root Cause :
Blocks from N nodes may not arrive in order. For example Node A found Block 1 on top of Block 0, and Node B found Block 2 on top of Block 1. Node C may receive Block 2 first, and then receive block 1.

Solution :
Need to keep a consensual block as an orphan block if the previous block is not the tip of the chain yet.
Also, whenever a consensual block is attached to the tip of the block, need to RECURSIVELY check if there is any orphan block whose parent is the tip of the chain.