Need to generate raw transactions deterministically.
Why

We need to run performance tests over and over, but generating transaction data takes long time.

How

Configure 'mining address' of '_FOR_TEST_ONLY' account for mining in scalechain.conf
import the mining address to the _FOR_TEST_ONLY account upon start up, if the getaccountaddress RPC does not return the given address.
Imeplement 'generaterawtransactions' RPC with three parameters. (1) a private key, which is the private key of the above 'mining address'. Necesssary to sign transactions (2) count, which describes the number of transaction groups that can be sent simultaneously (3) the number of transactions in a group.

Issue : #99