Node recovery using PBFT
What
Need to recover a node if it did not receive recent ledger data.

Why
Need to catch up other nodes by node recovery to get updates of the latest ledger data.

How
Calculate the max height of the blockchain of a peer from Version message sent by each peer. If the max height of the blockchain is different from local one, start IBD(Initial block download process).
The peer which sent the max height in the Version message becomes 'sender' of IBD.
Send getblocks message with a summary of local blockchain to 'sender'
For each inv message from the 'sender', request 'block' message if the block does not exist. For block message received from the 'sender', check previous block hash of the received block. If the previous block hash matches the best block hash of local blockchain, put the block on top of the best block.
After catching up to the max height of the sender blockchain, send "getblocks" message once more to the 'sender'.
Repeat the same steps at 4.
While Step 5,6 was being executed, we will receive consensus headers and blocks from other nodes. If we have a consensual block to put on our local blockchain, we are ready to be in consensus with other nodes. Put the block into the local blockchain, and stop IBD process.

Issue #1