Fix issue : There is a possibility that duplicate transactions exist in a block.
- Problem
There is a possibility that duplicate transactions exist in a block.

- Root cause
When transactions are put into the transaction pool, a the transaction time index whose key is (transaction time, transaction id) creates a key.

When the same transaction is put into the transaction pool by two threads, the index can have the same transaction id twice with different transaction time.

- Solution
1. Using fine grained locking based on the transaction hash, synchronise each transaction when it is attached.
2. Use the following library :
https://github.com/mojgh/JKeyLockManager

Issue : #105