null pointer exception in BlockChain.putBlock issue-#129
problem:
For an orphan block, the parent block does not exist yet, but the following code in BlockChain.putBlock assumes that the parent block exists.

val blockInfo = storage.getBlockInfo(db, blockHash)!!

root cause:

BlockChain.putBlock assumes that the parent block exists.

solution:
In BlockProcessor.acceptBlock, check if the parent block of the new block exists. If it does not exist, put it into the orphanage.

repro:
./run-prepare.sh
./run-multi.sh

issue link:
https://github.com/ScaleChain/scalechain/issues/129