Fix for Issue #130 Memory Leak ByteBuf.release() was not called (Netty)
Problem : Memory leaks while unit testing and e2e testing.
Root cause : ByteBuf was allocated inside of netty, but release was not called after using the buffer.
Solution : call release after the ByteBuf was used. If the size of the buffer is zero, no need to call release, as the retention count does not decrease.